requirements.txt
pystray
requests
yandex_music
pillow
pypresence
selenium==4.8.1
wxPython